improve types exports
