+++ README.md
@@ -67,7 +67,6 @@ const factory = new IExecOracleFactory(signer);
 ## Documentation
 
 - [OracleFactory documentation](https://tools.docs.iex.ec/tools/oracle-factory)
-<!-- - [OracleFactory technical design](./technical-design/index.md) -->
 - [iExec Protocol documentation](https://protocol.docs.iex.ec)
 
 ## License
