+++ src/index.ts
@@ -1,7 +1,7 @@
-export * from './types/public-types.js';
+export type * from './types/public-types.js';
+
 export { IExecOracleFactory } from './oracleFactory/OracleFactory.js';
 export { IExecOracleReader } from './oracleFactory/OracleReader.js';
 export * as utils from './utils/utils.js';
 export * as errors from './utils/errors.js';
-
 export { getWeb3Provider } from './utils/getWeb3Provider.js';
